Add parameterized CFN stack factory fixture
The test_networking and test_api_infrastructure integration tests
validate CloudFormation templates, and need to create a stack and run
validation on its resources and outputs.

This patch introduces a fixture that returns a parameterized
CloudFormation stack factory and manages the CloudFormation stack
creation and deletion.
